requirements.txt
# YouTube動画ダウンロード
yt-dlp>=2024.8.6

# 音声・動画処理
moviepy==1.0.3
pydub>=0.25.1
curl-cffi

# 音声認識・翻訳
openai-whisper>=20231117
google-cloud-translate>=3.15.3
googletrans>=4.0.0rc1

# 字幕処理
pysrt>=1.1.2
webvtt-py>=0.4.6

# Web framework（将来のUI用）
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# ユーティリティ
python-dotenv>=1.0.0
requests>=2.32.2
tqdm>=4.66.1
colorlog>=6.8.0

# AI関連（将来の機能用）
openai>=1.3.8
anthropic>=0.34.0

# データベース（設定保存用）
# sqlite3は標準ライブラリなので不要
